refactor: Move parser unit tests to decode_test and delete parser.go files
Code moved to packages: tags.go, playlist.go and decode.go
